% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/params.R
\name{charlistopts}
\alias{charlistopts}
\title{Change list version of paramaters into char vector}
\usage{
charlistopts(x)
}
\arguments{
\item{x}{A list.}
}
\value{
A character vector

Not expected to be used directly.
}
\description{
Turns a list of options into a charvar of options
}
\details{
The resulting charvar has values such as "name=value" where "name" was the named item
of the list.
}

