% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\name{NHumbersidecas}
\alias{NHumbersidecas}
\title{North Humberside leukemian and lymphoma example-- cases}
\format{A data frame with 191 observations and 2 variables}
\source{
Distributed with SaTScan software: \url{http://www.satscan.org}
}
\description{
A data set from North Humberside.  The variables are as follows:
}
\details{
\itemize{
  \item locationid:     Postal code ID
  \item numcases:        The number of cases observed
  }
}

