% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\name{NHumbersidegeo}
\alias{NHumbersidegeo}
\title{North Humberside leukemian and lymphoma example-- geography}
\format{A data frame with 191 observations and 3 variables}
\source{
Distributed with SaTScan software: \url{http://www.satscan.org}
}
\description{
A data set from North Humberside.  The variables are as follows:
}
\details{
\itemize{
  \item locationid:     Postal code ID
  \item x-coordinate:     x-coordinate
  \item y-coordinate:     y-coordinate
  }
}

