% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\name{NMcas}
\alias{NMcas}
\title{New Mexico Brain Cancer example-- cases}
\format{A data frame with 1175 observations and 5 variables}
\source{
Distributed with SaTScan software: \url{http://www.satscan.org}
}
\description{
A data set from New Mexico.  The variables are as follows:
}
\details{
\itemize{
  \item county:     County name
  \item cases:     Number of cases
  \item year:     Year of case
  \item agegroup:     Age group of case
  \item sex:      Sex of case
  }
}

