% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\name{NMgeo}
\alias{NMgeo}
\title{New Mexico Brain Cancer example-- geography}
\format{A data frame with 194 observations and 3 variables}
\source{
Distributed with SaTScan software: \url{http://www.satscan.org}
}
\description{
A data set from New Mexico.  The variables are as follows:
}
\details{
\itemize{
  \item county:     The US Postal Service ZIP code
  \item cases:   The number of cases observed
  \item long:    The date on which the cases were observed in that ZIP code
  }
}

