% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\name{NMpop}
\alias{NMpop}
\title{New Mexico Brain Cancer example-- population}
\format{A data frame with 3456 observations and 5 variables}
\source{
Distributed with SaTScan software: \url{http://www.satscan.org}
}
\description{
A data set from New Mexico.  The variables are as follows:
}
\details{
\itemize{
  \item county:     County name
  \item year:     year
  \item year:     population
  \item agegroup:     Age group
  \item sex:      Sex
  }
}

