% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DataDoc.R
\name{NYCfevergeo}
\alias{NYCfevergeo}
\title{New York City Fever example-- geography}
\format{A data frame with 192 observations and 3 variables}
\source{
Distributed with SaTScan software: \url{http://www.satscan.org}
}
\description{
A data set from New York City.  The variables are as follows:
}
\details{
\itemize{
  \item zip:     The US Postal Service ZIP code
  \item lat:     Decimal latitude north
  \item long:    Decimal longitude
  }
}

