% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods.r
\name{print.satscan}
\alias{print.satscan}
\title{Methods for satscan-class objects}
\usage{
\method{print}{satscan}(x, ...)
}
\arguments{
\item{x}{is a satscan object}

\item{...}{vestigial, for compatability with the default summary method}
}
\value{
x, invisibly.  Side effect is to display ss$main, the SaTScan text report
}
\description{
These functions define the default methods for satscan-class objects, which
are the result objects from a call to satscan()
}

