% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readfiles.R
\name{read.llr}
\alias{read.llr}
\title{Read SaTScan output files}
\usage{
read.llr(location, file)
}
\arguments{
\item{location}{A directory location, including the trailing "/"}

\item{file}{A file name, without the extension.}
}
\value{
A data frame.
}
\description{
Reads a SaTScan output .dbf file.
}
\details{
This is expected to be a purely internal function.
It's called by \code{satscan()} with the location and file nmae provided
to that function.  Since it's nothing more than \code{foreign::read.dbf()}, it's
probably nor necessary to even have it as a function.
}

