% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{search_for_param}
\alias{search_for_param}
\title{find earliest SaTScan parameter set that contains specified parameter}
\usage{
search_for_param(ssenv, param)
}
\arguments{
\item{ssenv}{the SaTScan environment to search for the specified parameter in}

\item{param}{the parameter to search for in the SaTScan environment}
}
\value{
A string specifying the earliest parameter set that contains the specified
parameter. If the parameter is not found in any parameter set, then an empty string
is returned.
}
\description{
This function sorts the SaTScan parameter sets by version number.
Then, it searches forward from the current version to find which (if any)
parameter sets contain the specified parameter. If any parameter set contains
the specified parameter, then that parameter set's version number is returned 
as a string.

Not expected to be used directly.
}
