% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/params.R
\name{subin}
\alias{subin}
\title{Substitute new values into the input object}
\usage{
subin(x, ssparams)
}
\arguments{
\item{x}{A character vector of the form "name=value"}

\item{ssparams}{A character vector with arbitrary lines,
currently imagined to be .ss.params}
}
\value{
The modified ssparams.
}
\description{
Replaces existing values found in one object with new values
}
\details{
For each line of x, the function: 1) finds the "name" and the "value"
2) checks to see whether the "name" exists in ssparams; if not, prints a warning
but if so, replaces the existing line of ssparams with that line of x.

Not expected to be used directly.
}

