% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods.r
\name{summary.satscan}
\alias{summary.satscan}
\title{Methods for satscan-class objects}
\usage{
\method{summary}{satscan}(object, ...)
}
\arguments{
\item{object}{is a satscan object}

\item{...}{vestigial, for compatability with the default summary method}
}
\value{
object, invisibly.  Side effect is to display minimal facts contained in ss
}
\description{
These functions define the default methods for satscan-class objects, which
are the result objects from a call to satscan()
}

