% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{subin}
\alias{subin}
\title{Substitute new values into the input object}
\usage{
subin(x, ssparams)
}
\arguments{
\item{x}{A character vector of the form "name=value"}

\item{ssparams}{A character vector with arbitrary lines, 
currently imagined to be .ss.params}
}
\value{
The modified ssparams.
}
\description{
Replaces existing values found in one object with new values
}
\details{
For each line of x, the function: 1) finds the "name" and the "value"
2) checks to see whether the "name" exists in ssparams.
If the "name" exists in .ss.params, then the existing line is replaced
with that line of x
If the "name" does not exist in .ss.params, then later parameter sets
are check to see if the "name" exists in them. If the "name" exists
in a later parameter set, this is printed as a note to the user. If 
the "name" is not found in any parameter set, then a warning is given.

Not expected to be used directly.
}
